frontend/requirements.txt
Flask==2.3.2
rasterio==1.3.3
Pillow==9.4.0
numpy==1.24.2
requests